Extract correct library version for Python 3.6 on Linux
Apparently, the version on Linux uses 3.6m as the version strings which
was not extracted previously.

This *should* not create a regression because there will most likely be
a dot (.) or other separator after the version number even in the case
that the shared library extension or anything else that might come after
the version number contains an "m".